cmv_ml/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
python-dotenv
alembic
cryptography
python-jose
passlib[bcrypt]
bcrypt==4.1.3
httpx
xgboost
joblib
shap
numpy
hypothesis
pytest
pytest-asyncio
pytest-mock
